thank you Cody's wife

+++ doubly_linked_list/doubly_linked_list.py
@@ -12,15 +12,25 @@ class ListNode:
         return self.value
 
     def get_next(self):
-        return self.next
+        if self.next:
+            return self.next
+        else:
+            return None
 
     def get_previous(self):
-        return self.prev
+        if self.prev:
+            return self.prev
+        else:
+            return None
     
     def delete(self):
-        self.prev = None
-        self.next = None
+        if self.prev:
+            self.prev.next = self.next
+        if self.next:
+            self.next.prev = self.prev
         self.value = None
+        self.next = None
+        self.prev = None
 """
 Our doubly-linked list class. It holds references to 
 the list's head and tail nodes.
@@ -134,15 +144,35 @@ class DoublyLinkedList:
     order of the other elements of the List.
     """
     def delete(self, node):
-        nextNode = node.get_next()
-        previousNode = node.get_previous()
-        previousNode.next = nextNode
-        nextNode.prev = previousNode
-        node.delete()
+        if not self.head and not self.tail:
+            return
+
+        if self.head == self.tail:
+            self.head = None
+            self.tail = None
+        elif self.head == node:
+            self.head = node.get_next()
+            node.delete()
+        elif self.tail == node:
+            self.tail == node.get_previous()
+            node.delete()
+        else:
+            node.delete()
+        self.length += 1
+
 
     """
     Finds and returns the maximum value of all the nodes 
     in the List.
     """
     def get_max(self):
-        pass
+        if not self.head:
+            return None
+        max_value = self.head.value
+        current_node = self.head
+        while current_node:
+            if current_node.value > max_value:
+                max_value = current_node.value
+            current_node = current_node.get_next()
+
+        return max_value
